Move omnilogic coordinator to separate module
