+++ homeassistant/components/omnilogic/__init__.py
@@ -10,7 +10,6 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
 from homeassistant.helpers import aiohttp_client
 
-from .common import OmniLogicUpdateCoordinator
 from .const import (
     CONF_SCAN_INTERVAL,
     COORDINATOR,
@@ -18,6 +17,7 @@ from .const import (
     DOMAIN,
     OMNI_API,
 )
+from .coordinator import OmniLogicUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
