+++ homeassistant/components/omnilogic/sensor.py
@@ -15,8 +15,9 @@ from homeassistant.const import (
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .common import OmniLogicEntity, OmniLogicUpdateCoordinator, check_guard
+from .common import OmniLogicEntity, check_guard
 from .const import COORDINATOR, DEFAULT_PH_OFFSET, DOMAIN, PUMP_TYPES
+from .coordinator import OmniLogicUpdateCoordinator
 
 
 async def async_setup_entry(
